Refactor package names to follow lowercase and simplified naming conventions
This change standardizes package naming across the UI layer to improve consistency and adhere to standard Kotlin and Android package naming guidelines.

Included changes:
- Renamed `ui.deviceInfo` to `ui.deviceinfo` and updated all associated ViewModels and Composables.
- Renamed `ui.remoteControl` and its subpackages to `ui.remotecontrol` across both the main UI and settings modules.
- Simplified EPG package names by shortening `ui.epg.tvEpg`, `ui.epg.radioEpg`, and `ui.epg.serviceEpg` to `ui.epg.tv`, `ui.epg.radio`, and `ui.epg.service` respectively.
- Updated all internal file package declarations and imports, including references within `NavigationDisplay.kt`, to reflect the new directory structure.
