+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/deviceinfo/DeviceInfoViewModel.kt
@@ -17,7 +17,7 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.ui.deviceInfo
+package io.github.deprec8.enigmadroid.ui.deviceinfo
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
