+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/epg/radio/RadioEpg.kt
@@ -17,7 +17,7 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.ui.epg.radioEpg
+package io.github.deprec8.enigmadroid.ui.epg.radio
 
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
