+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/epg/service/ServiceEpgViewModel.kt
@@ -17,22 +17,20 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.ui.epg.tvEpg
+package io.github.deprec8.enigmadroid.ui.epg.service
 
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import io.github.deprec8.enigmadroid.common.enums.ContentType
 import io.github.deprec8.enigmadroid.common.enums.LoadingState
 import io.github.deprec8.enigmadroid.data.ApiRepository
 import io.github.deprec8.enigmadroid.data.DevicesRepository
 import io.github.deprec8.enigmadroid.data.LoadingRepository
 import io.github.deprec8.enigmadroid.data.SearchHistoryRepository
 import io.github.deprec8.enigmadroid.data.SettingsRepository
-import io.github.deprec8.enigmadroid.model.api.Bouquet
 import io.github.deprec8.enigmadroid.model.api.Event
-import io.github.deprec8.enigmadroid.model.api.EventBatchSet
+import io.github.deprec8.enigmadroid.model.api.EventBatch
 import io.github.deprec8.enigmadroid.model.api.search
 import io.github.deprec8.enigmadroid.ui.components.search.asHighlightedWords
 import kotlinx.coroutines.Job
@@ -48,7 +46,7 @@ import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 @HiltViewModel
-class TvEpgViewModel @Inject constructor(
+class ServiceEpgViewModel @Inject constructor(
     private val apiRepository: ApiRepository,
     private val loadingRepository: LoadingRepository,
     private val searchHistoryRepository: SearchHistoryRepository,
@@ -56,23 +54,17 @@ class TvEpgViewModel @Inject constructor(
     private val devicesRepository: DevicesRepository
 ) : ViewModel() {
 
-    private val _eventBatchSet = MutableStateFlow<EventBatchSet?>(null)
-    val eventBatchSet: StateFlow<EventBatchSet?> = _eventBatchSet.asStateFlow()
-
-    private val _filteredEvents = MutableStateFlow<List<Event>?>(null)
-    val filteredEvents: StateFlow<List<Event>?> = _filteredEvents.asStateFlow()
+    private val _eventBatch = MutableStateFlow<EventBatch?>(null)
+    val eventBatch: StateFlow<EventBatch?> = _eventBatch.asStateFlow()
 
     private val _loadingState = MutableStateFlow(LoadingState.LOADING)
     val loadingState: StateFlow<LoadingState> = _loadingState.asStateFlow()
 
+    private val _filteredEvents = MutableStateFlow<List<Event>?>(null)
+    val filteredEvents: StateFlow<List<Event>?> = _filteredEvents.asStateFlow()
+
     private val _searchHistory = MutableStateFlow<List<String>>(emptyList())
     val searchHistory: StateFlow<List<String>> = _searchHistory.asStateFlow()
-
-    private val _currentBouquetReference = MutableStateFlow("")
-    val currentBouquetReference: StateFlow<String> = _currentBouquetReference.asStateFlow()
-
-    private val _bouquets = MutableStateFlow<List<Bouquet>?>(null)
-    val bouquets: StateFlow<List<Bouquet>?> = _bouquets.asStateFlow()
 
     val searchFieldState = TextFieldState()
 
@@ -88,7 +80,7 @@ class TvEpgViewModel @Inject constructor(
 
     private var fetchJob: Job? = null
 
-    private var fetchedEventBatchSetMap = emptyMap<String, EventBatchSet>()
+    private var serviceReference = ""
 
     private var loadedDeviceId: Int? = null
 
@@ -99,14 +91,14 @@ class TvEpgViewModel @Inject constructor(
             }
         }
         viewModelScope.launch {
-            combine(_eventBatchSet, searchInput) { eventBatchSet, searchInput ->
-                eventBatchSet?.eventBatches?.flatMap { it.events }?.search(searchInput)
+            combine(_eventBatch, searchInput) { eventBatch, searchInput ->
+                eventBatch?.events?.search(searchInput)
             }.collectLatest {
                 _filteredEvents.value = it
             }
         }
         viewModelScope.launch {
-            searchHistoryRepository.getTvEpgSearchHistory().collectLatest {
+            searchHistoryRepository.getServiceEpgSearchHistory().collectLatest {
                 _searchHistory.value = it
             }
         }
@@ -117,6 +109,10 @@ class TvEpgViewModel @Inject constructor(
         }
     }
 
+    fun initialize(serviceReference: String) {
+        this.serviceReference = serviceReference
+    }
+
     suspend fun updateLoadingState(isForcedUpdate: Boolean) {
         loadingRepository.updateLoadingState(isForcedUpdate)
     }
@@ -125,40 +121,17 @@ class TvEpgViewModel @Inject constructor(
         viewModelScope.launch {
             val currentDeviceId = devicesRepository.getCurrentDeviceId().first()
             if (currentDeviceId != loadedDeviceId || isForced) {
-                _eventBatchSet.value = null
-                _bouquets.value = null
-                fetchedEventBatchSetMap = emptyMap()
+                _eventBatch.value = null
                 loadedDeviceId = currentDeviceId
             }
 
-            if (_eventBatchSet.value == null || _bouquets.value == null) {
+            if (_eventBatch.value == null) {
                 fetchJob?.cancel()
                 fetchJob = launch {
-                    fetchBouquets()
-                    fetchEpgBatchSet()
+                    _eventBatch.value = apiRepository.fetchServiceEpgBatch(serviceReference)
                 }
             }
         }
-    }
-
-    private suspend fun fetchBouquets() {
-        val bouquets = apiRepository.fetchBouquets(ContentType.Tv)
-        _bouquets.value = bouquets
-        val firstBRef = bouquets.firstOrNull()?.reference ?: ""
-        if (_currentBouquetReference.value.isBlank() || bouquets.find { it.reference == _currentBouquetReference.value } == null) {
-            _currentBouquetReference.value = firstBRef
-        }
-    }
-
-    private suspend fun fetchEpgBatchSet() {
-        val cached = fetchedEventBatchSetMap[_currentBouquetReference.value]
-        if (cached != null) {
-            _eventBatchSet.value = cached
-            return
-        }
-        val epgBatchSet = apiRepository.fetchEpgEventBatchSet(_currentBouquetReference.value)
-        _eventBatchSet.value = epgBatchSet
-        fetchedEventBatchSetMap += _currentBouquetReference.value to epgBatchSet
     }
 
     fun addTimerForEvent(event: Event) {
@@ -169,20 +142,11 @@ class TvEpgViewModel @Inject constructor(
         }
     }
 
-    fun setCurrentBouquet(bouquetReference: String) {
-        fetchJob?.cancel()
-        _currentBouquetReference.value = bouquetReference
-        _eventBatchSet.value = null
-        fetchJob = viewModelScope.launch {
-            fetchEpgBatchSet()
-        }
-    }
-
     fun updateSearchInput() {
         val input = searchFieldState.text.toString()
         if (input.isNotBlank()) {
             viewModelScope.launch {
-                searchHistoryRepository.addToTvEpgSearchHistory(input)
+                searchHistoryRepository.addToServiceEpgSearchHistory(input)
             }
         }
         searchInput.value = input
