+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/main/components/NavigationDisplay.kt
@@ -46,22 +46,22 @@ import io.github.deprec8.enigmadroid.ui.components.navigation.LocalSharedTransit
 import io.github.deprec8.enigmadroid.ui.components.navigation.Navigator
 import io.github.deprec8.enigmadroid.ui.components.navigation.toEntries
 import io.github.deprec8.enigmadroid.ui.current.CurrentPage
-import io.github.deprec8.enigmadroid.ui.deviceInfo.DeviceInfoPage
-import io.github.deprec8.enigmadroid.ui.epg.radioEpg.RadioEpgPage
-import io.github.deprec8.enigmadroid.ui.epg.serviceEpg.ServiceEpgPage
-import io.github.deprec8.enigmadroid.ui.epg.serviceEpg.ServiceEpgViewModel
-import io.github.deprec8.enigmadroid.ui.epg.tvEpg.TvEpgPage
+import io.github.deprec8.enigmadroid.ui.deviceinfo.DeviceInfoPage
+import io.github.deprec8.enigmadroid.ui.epg.radio.RadioEpgPage
+import io.github.deprec8.enigmadroid.ui.epg.service.ServiceEpgPage
+import io.github.deprec8.enigmadroid.ui.epg.service.ServiceEpgViewModel
+import io.github.deprec8.enigmadroid.ui.epg.tv.TvEpgPage
 import io.github.deprec8.enigmadroid.ui.live.radio.RadioPage
 import io.github.deprec8.enigmadroid.ui.live.tv.TvPage
 import io.github.deprec8.enigmadroid.ui.movies.MoviesDirectoryPage
 import io.github.deprec8.enigmadroid.ui.movies.MoviesPage
 import io.github.deprec8.enigmadroid.ui.movies.MoviesViewModel
-import io.github.deprec8.enigmadroid.ui.remoteControl.RemoteControlPage
+import io.github.deprec8.enigmadroid.ui.remotecontrol.RemoteControlPage
 import io.github.deprec8.enigmadroid.ui.settings.SettingsPage
 import io.github.deprec8.enigmadroid.ui.settings.about.AboutPage
 import io.github.deprec8.enigmadroid.ui.settings.about.LibrariesPage
 import io.github.deprec8.enigmadroid.ui.settings.devices.DevicesPage
-import io.github.deprec8.enigmadroid.ui.settings.remoteControl.RemoteControlSettingsPage
+import io.github.deprec8.enigmadroid.ui.settings.remotecontrol.RemoteControlSettingsPage
 import io.github.deprec8.enigmadroid.ui.settings.search.SearchSettingsPage
 import io.github.deprec8.enigmadroid.ui.signal.SignalPage
 import io.github.deprec8.enigmadroid.ui.timers.TimersPage
