+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/remotecontrol/RemoteControlViewModel.kt
@@ -17,7 +17,7 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.ui.remoteControl
+package io.github.deprec8.enigmadroid.ui.remotecontrol
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
