+++ app/src/main/java/io/github/deprec8/enigmadroid/ui/remotecontrol/components/ControlButtons.kt
@@ -17,7 +17,7 @@
  * along with EnigmaDroid.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package io.github.deprec8.enigmadroid.ui.remoteControl.components
+package io.github.deprec8.enigmadroid.ui.remotecontrol.components
 
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.widthIn
